Using a modular config
