+++ userportal/settings/10-base.conf
@@ -1,6 +1,20 @@
+"""
+Django settings for userportal project.
+
+Generated by 'django-admin startproject' using Django 3.1.7.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/3.1/topics/settings/
+
+For the full list of settings and their values, see
+https://docs.djangoproject.com/en/3.1/ref/settings/
+"""
+
 from pathlib import Path
 from django.utils.translation import gettext_lazy as _
 import os
+import saml2
+import saml2.saml
 import logging.config
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
@@ -14,7 +28,7 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 SECRET_KEY = 'changeme'
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
+DEBUG = False
 
 ALLOWED_HOSTS = ['localhost']
 
@@ -28,7 +42,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'django.contrib.humanize',
-    'freeipa_auth',
+    'djangosaml2',
     'watchman',
 
     'pages',
@@ -36,7 +50,13 @@ INSTALLED_APPS = [
     'notes',
 
     'jobstats',
+    #'accountstats',
+    #'cloudstats',
+    # 'quotas',
     'top',
+    #'usersummary',
+
+    'ccldap',
 
     # system modules
     'rest_framework',
@@ -45,9 +65,6 @@ INSTALLED_APPS = [
 ]
 
 MIDDLEWARE = []
-if DEBUG:
-    INSTALLED_APPS += ['debug_toolbar']
-    MIDDLEWARE = ['debug_toolbar.middleware.DebugToolbarMiddleware']
 
 MIDDLEWARE += [
     'django.middleware.security.SecurityMiddleware',
@@ -87,39 +104,6 @@ STATICFILES_DIRS = [
 ]
 
 WSGI_APPLICATION = 'userportal.wsgi.application'
-
-
-# Database
-# https://docs.djangoproject.com/en/3.1/ref/settings/#databases
-
-DATABASES = {
-    'default': {
-        'ENGINE': 'django.db.backends.mysql',
-        'NAME': 'userportal',
-        'USER': 'userportal',
-        'PASSWORD': 'changeme',
-        'HOST': 'changeme',
-        'PORT': '3306',
-    },
-    'slurm': {
-        'ENGINE': 'django.db.backends.mysql',
-        'NAME': 'slurmdb',
-        'USER': 'slurm-stats',
-        'PASSWORD': 'changeme',
-        'HOST': 'changeme',
-        'PORT': '3306',
-    },
-}
-
-WATCHMAN_DATABASES = ['default', 'slurm']
-
-DATABASE_ROUTERS = ['database_routers.dbrouters.DbRouter']
-
-PROMETHEUS = {
-    'url': 'http://localhost:9090',
-    'headers': {},
-    'filter': "",
-}
 
 # Password validation
 # https://docs.djangoproject.com/en/3.1/ref/settings/#auth-password-validators
@@ -162,67 +146,19 @@ LOCALE_PATHS = (
     os.path.join(BASE_DIR, 'locale'),
 )
 
+# Static files (CSS, JavaScript, Images)
+# https://docs.djangoproject.com/en/3.1/howto/static-files/
+
 STATIC_URL = '/static/'
-STATIC_ROOT = '/var/www/userportal-static/'
+STATIC_ROOT = '/opt/userportal/collected-static/'
 
 AUTHENTICATION_BACKENDS = [
     'userportal.authentication.staffRemoteUserBackend',
     'django.contrib.auth.backends.ModelBackend',
-    'freeipa_auth.backends.FreeIpaRpcAuthBackend',
-    ]
-
-FREEIPA_AUTH_BACKEND_ENABLED = True
-FREEIPA_AUTH_SERVER = "changeme" # defaults to None
-FREEIPA_AUTH_SSL_VERIFY = False# FIXME # this would be the path to the ssl cert used
-FREEIPA_AUTH_UPDATE_USER_GROUPS = True
-FREEIPA_AUTH_ALWAYS_UPDATE_USER = True
-FREEIPA_AUTH_USER_ATTRS_MAP = {"first_name": "givenname", "last_name": "sn", "email": "mail"}
-FREEIPA_AUTH_SERVER_TIMEOUT = 5
-FREEIPA_AUTH_STAFF_GROUPS = ['admins']
-
-LOGIN_REDIRECT_URL = '/'
-
-EXPORTER_INSTALLED = [
-    'slurm-job-exporter',
-    'node_exporter',
+    'userportal.authentication.staffSaml2Backend',
 ]
 
-CLUSTER_NAME = 'cluster'  # used for slurm DB name
-LUSTRE_FS_NAMES = None
-LOGINS = {
-    'login1': {
-        'network_interface': 'eth0',
-    },
-}
-DTNS = None
-
-AMORTIZATION_YEARS = None
-CPU_CORE_COST_PER_HOUR = None
-GPU_COST_PER_HOUR = None
-
-ELECTRICITY_COST_PER_KWH = None
-COOLING_COST_PER_KWH = None
-CO2_KG_PER_MWH = None
-ELECTRIC_CAR_RANGE_KM_PER_KWH = None
-
-CLOUD_CPU_CORE_COST_PER_HOUR = None
-CLOUD_GPU_COST_PER_HOUR = None
-
-# Used to set reasonable thresholds for the mem usage in the "top compute" page
-NORMAL_MEM_BY_CORE = 256/64 * 1024 * 1024 * 1024  # 256GB for 64 cores
-NORMAL_MEM_BY_GPU = 512 / 4 * 1024 * 1024 * 1024  # 512GB for 4 GPUs
-NORMAL_CORES_BY_GPU = 48 / 4  # 48 cores for 4 GPUs
-
-SLURM_TRES = {
-    '1001=': 'NVIDIA A100-SXM4-40GB',
-}
-
-CLUSTER_NAME_TITLE = 'MagicCastle'
-FAVICON = 'https://object-arbutus.cloud.computecanada.ca/userportal-public/narval.png'
-
-CLUSTER_INTRO = """
-<p>Welcome to this MagicCastle cluster</p>
-"""
+LOGIN_REDIRECT_URL = '/'
 
 SETTINGS_EXPORT = [
     'CLUSTER_NAME_TITLE',
