+++ userportal/settings/21-prometheus.conf
@@ -1,0 +1,5 @@
+PROMETHEUS = {
+    'url': 'https://thanos.dant.computecanada.ca',
+    'headers': {'Authorization': 'Basic changeme_base64'},
+    'filter': "cluster='narval'",
+}
